+++ adminapp/src/pages/MemberForm.jsx
@@ -3,25 +3,12 @@ import AddressInputs from "../components/AddressInputs";
 import AutocompleteSearch from "../components/AutocompleteSearch";
 import FormLayout from "../components/FormLayout";
 import ResponsiveStack from "../components/ResponsiveStack";
-import { useGlobalApiState } from "../hooks/globalApiState";
-import useRoleAccess from "../hooks/useRoleAccess";
+import RoleEditor from "../components/RoleEditor";
 import mergeAt from "../shared/mergeAt";
 import withoutAt from "../shared/withoutAt";
 import AddIcon from "@mui/icons-material/Add";
-import AddCircleOutlineIcon from "@mui/icons-material/AddCircleOutline";
 import DeleteIcon from "@mui/icons-material/Delete";
-import {
-  Box,
-  Chip,
-  CircularProgress,
-  Divider,
-  FormHelperText,
-  FormLabel,
-  Icon,
-  Stack,
-  TextField,
-  Typography,
-} from "@mui/material";
+import { Box, Divider, FormLabel, Icon, Stack, TextField } from "@mui/material";
 import Button from "@mui/material/Button";
 import merge from "lodash/merge";
 import React from "react";
@@ -75,7 +62,7 @@ export default function MemberForm({
           helperText="11-digit US phone number, begins with 1."
           onChange={setFieldFromInput}
         />
-        <Roles roles={resource.roles} setRoles={(r) => setField("roles", r)} />
+        <RoleEditor roles={resource.roles} setRoles={(r) => setField("roles", r)} />
         <Divider />
         <FormLabel>Legal Entity</FormLabel>
         <AddressInputs
@@ -95,59 +82,6 @@ export default function MemberForm({
   );
 }
 
-function Roles({ roles, setRoles }) {
-  const { canWriteResource } = useRoleAccess();
-  const allRoles = useGlobalApiState(api.getRoles, null, { pick: (r) => r.data.items });
-
-  if (!canWriteResource("role")) {
-    return null;
-  }
-
-  function deleteRole(id) {
-    const newRoles = roles.filter((c) => c.id !== id);
-    setRoles(newRoles);
-  }
-
-  function handleAdd(newRole) {
-    const newRoles = [...roles, newRole];
-    setRoles(newRoles);
-  }
-
-  const hasRoleIds = new Set();
-  roles.forEach((r) => hasRoleIds.add(r.id));
-
-  return (
-    <Box>
-      <FormLabel>Roles</FormLabel>
-      <FormHelperText>
-        If you remove special roles like "admin", you will be logged out of this account.
-      </FormHelperText>
-      <Stack direction="row" gap={1} sx={{ marginTop: 1, flexWrap: "wrap" }}>
-        {allRoles === null && <CircularProgress />}
-        {allRoles?.map((r) => {
-          const hasRole = hasRoleIds.has(r.id);
-          const handler = hasRole ? () => deleteRole(r.id) : () => handleAdd(r);
-          return (
-            <Chip
-              key={r.id}
-              label={r.label}
-              color={hasRole ? "success" : undefined}
-              title={hasRole ? "Delete Role" : "Add Role"}
-              deleteIcon={hasRole ? null : <AddCircleOutlineIcon />}
-              onClick={handler}
-              onDelete={handler}
-            />
-          );
-        })}
-        {allRoles && allRoles.length === 0 && (
-          <Typography>
-            * No roles available, ask developers for help if you see this
-          </Typography>
-        )}
-      </Stack>
-    </Box>
-  );
-}
 function OrganizationMemberships({ memberships, setMemberships, memberId }) {
   // Include member id to associate with a new membership
   const initialOrganizationMembership = {
