+++ adminapp/src/pages/OrganizationDetailPage.jsx
@@ -1,11 +1,13 @@
 import api from "../api";
 import AdminLink from "../components/AdminLink";
+import AuditActivityList from "../components/AuditActivityList";
 import ProgramEnrollmentRelatedList from "../components/ProgramEnrollmentRelatedList";
 import RelatedList from "../components/RelatedList";
 import ResourceDetail from "../components/ResourceDetail";
 import { dayjs } from "../modules/dayConfig";
 import formatDate from "../modules/formatDate";
 import createRelativeUrl from "../shared/createRelativeUrl";
+import { Chip } from "@mui/material";
 import React from "react";
 
 export default function OrganizationDetailPage() {
@@ -19,6 +21,13 @@ export default function OrganizationDetailPage() {
         { label: "Created At", value: dayjs(model.createdAt) },
         { label: "Updated At", value: dayjs(model.updatedAt) },
         { label: "Name", value: model.name },
+        {
+          label: "Roles",
+          children: model.roles.map((role) => (
+            <Chip key={role.id} label={role.label} sx={{ mr: 0.5 }} />
+          )),
+          hideEmpty: true,
+        },
       ]}
     >
       {(model) => (
@@ -31,6 +40,7 @@ export default function OrganizationDetailPage() {
           <RelatedList
             title={`Memberships (${model.memberships.length})`}
             rows={model.memberships}
+            showMore
             addNewLabel="Create another membership"
             addNewLink={createRelativeUrl(`/membership/new`, {
               organizationId: model.id,
@@ -47,6 +57,7 @@ export default function OrganizationDetailPage() {
               formatDate(row.createdAt),
             ]}
           />
+          <AuditActivityList activities={model.auditActivities} />
         </>
       )}
     </ResourceDetail>
