+++ lib/suma/organization.rb
@@ -8,12 +8,15 @@ class Suma::Organization < Suma::Postgres::Model(:organizations)
   include Suma::AdminLinked
   include Suma::HasActivityAudit
 
+  plugin :association_pks
   plugin :hybrid_search
   plugin :timestamps
 
   one_to_many :memberships, class: "Suma::Organization::Membership", key: :verified_organization_id
   one_to_many :program_enrollments, class: "Suma::Program::Enrollment"
   many_to_many :roles, class: "Suma::Role", join_table: :roles_organizations
+
+  plugin :association_array_replacer, :roles
 
   def rel_admin_link = "/organization/#{self.id}"
